Refactored test file structure
Refactored test file structure to clearly separate unit tests and integration tests.
