+++ test/integration_test/bcccp/tickets/adhoc/Integration_AdhocTicketDAO_FactoryTicket.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package test;
+package integration_test.bcccp.tickets.adhoc;
 import bcccp.tickets.adhoc.*;
 import org.junit.After;
 import org.junit.AfterClass;
