+++ test/unit_test/bcccp/carpark/CarparkTest.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package bcccp.carpark;
+package unit_test.bcccp.carpark;
 
+import bcccp.carpark.Carpark;
+import bcccp.carpark.ICarparkObserver;
 import bcccp.tickets.adhoc.AdhocTicket;
 import bcccp.tickets.adhoc.AdhocTicketDAO;
 import bcccp.tickets.adhoc.IAdhocTicket;
