+++ test/unit_test/bcccp/tickets/adhoc/testAdhocTicketFactory.java
@@ -1,4 +1,4 @@
-package bcccp;
+package unit_test.bcccp.tickets.adhoc;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
@@ -7,6 +7,9 @@ import org.junit.*;
 import org.mockito.Mock;
 
 import bcccp.tickets.adhoc.*;
+import bcccp.tickets.adhoc.AdhocTicket;
+import bcccp.tickets.adhoc.AdhocTicketFactory;
+import bcccp.tickets.adhoc.IAdhocTicket;
 
 
 public class testAdhocTicketFactory {
