+++ test/unit_test/bcccp/tickets/season/UsageRecordTest.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package bcccp.tickets.season;
+package unit_test.bcccp.tickets.season;
 
+import bcccp.tickets.season.UsageRecord;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
